commit 2

+++ app/Http/Controllers/PostsController.php
@@ -7,6 +7,11 @@ use App\Post;
 
 class PostsController extends Controller
 {
+public function __construct()
+{
+    $this->middleware('auth',['except' =>['index','show']]);
+}
+
     /**
      * Display a listing of the resource.
      *
@@ -45,6 +50,7 @@ class PostsController extends Controller
         $post = new Post();
         $post->title =$request->input('title');
         $post->body =$request->input('body');
+        $post->user_id =auth()->user()->id;
         $post->save();
 
         return redirect('/posts')->with('success','Post Successfully Created!!!');
@@ -71,7 +77,14 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unautorized Page!!!');
+        }
+        
         return view('posts.edit')->with('post',$post);
+
+
     }
 
     /**
@@ -105,6 +118,11 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        if(auth()->user()->id != $post->user_id){
+            return redirect('/posts')->with('error','Unautorized Page!!!');
+        }
+
         $post->delete();
 
         return redirect('/posts')->with('success','Post Successfully Deleted!!!');
